fix: adjustment filter and modal
