+++ src/design/navigation/MenuItem/index.tsx
@@ -1,73 +1,36 @@
-import { Icon, Text, Stack } from "@inubekit/inubekit";
-import { basic } from "@design/tokens";
-import { MenuItemSpacingType } from "@ptypes/navigation/menuItem/ImenuItemSpacing";
-import { StyledMenuItemLink } from "./styles";
+import { IRenderMenuItems } from "@ptypes/navigation/IRenderMenuItems";
+import { MenuLink } from "../MenuLink";
+import { MenuOption } from "../MenuOption";
 
-interface IMenuItem {
-  title: string;
-  description?: string;
-  spacing?: MenuItemSpacingType;
-  iconBefore?: React.JSX.Element;
-  iconAfter?: React.JSX.Element;
-  isDisabled?: boolean;
-  path?: string;
-  onClick?: () => void;
-}
-
-const MenuItem = (props: IMenuItem) => {
-  const {
-    title,
-    description,
-    spacing = "wide",
-    iconBefore,
-    iconAfter,
-    isDisabled = false,
-    path = "#",
-    onClick,
-  } = props;
+const RenderMenuItems = (props: IRenderMenuItems) => {
+  const { options } = props;
   return (
-    <StyledMenuItemLink
-      spacing={spacing}
-      disabled={isDisabled}
-      to={path}
-      onClick={onClick}
-    >
-      <Stack gap={basic.spacing.s12} alignItems="center">
-        {iconBefore && (
-          <Icon
-            icon={iconBefore}
-            spacing="narrow"
-            size="24px"
-            appearance="dark"
-            disabled={isDisabled}
-          />
-        )}
-        <Stack direction="column" gap={basic.spacing.s4}>
-          <Text type="label" size="large" disabled={isDisabled}>
-            {title}
-          </Text>
-          <Text
-            type="body"
-            size="small"
-            appearance="gray"
-            disabled={isDisabled}
-          >
-            {description}
-          </Text>
-        </Stack>
-      </Stack>
-      {iconAfter && (
-        <Icon
-          icon={iconAfter}
-          spacing="narrow"
-          size="24px"
-          appearance="dark"
-          disabled={isDisabled}
-        />
-      )}
-    </StyledMenuItemLink>
+    <>
+      {options.map((option) => {
+        if (option.path) {
+          return (
+            <MenuLink
+              label={option.label}
+              key={option.id}
+              icon={option.icon}
+              path={option.path}
+            />
+          );
+        }
+        if (option.handleClick) {
+          return (
+            <MenuOption
+              label={option.label}
+              key={option.id}
+              icon={option.icon}
+              handleClick={option.handleClick}
+            />
+          );
+        }
+        return null;
+      })}
+    </>
   );
 };
 
-export { MenuItem };
-export type { IMenuItem };
+export { RenderMenuItems };
