+++ src/design/navigation/MenuSection/styles.ts
@@ -1,4 +1,4 @@
-import { IMenuItem } from "../MenuItem";
+import { IMenuItem } from "@ptypes/navigation/menuItem/IMenuItem";
 
 interface ISection {
   title?: string;
