+++ src/design/navigation/index.tsx
@@ -1,43 +1,11 @@
-import { useEffect, useRef } from "react";
-import { useCallback } from "react";
+import { useEffect, useRef, useCallback } from "react";
 import { Stack } from "@inubekit/inubekit";
-import { IOption } from "@ptypes/navigation/IOption";
+import { IMenu } from "@ptypes/navigation/IMenu";
 import { StyledMenu, StyledMenuContainer } from "./styles";
-import { MenuLink } from "./MenuLink";
-import { MenuOption } from "./MenuOption";
+import { RenderMenuItems } from "./MenuItem";
 
-interface IMenu {
-  options: IOption[];
-  handleClose: () => void;
-}
-
-const renderMenuItems = (options: IOption[]) => {
-  return options.map((option) => {
-    if (option.path) {
-      return (
-        <MenuLink
-          label={option.label}
-          key={option.id}
-          icon={option.icon}
-          path={option.path}
-        />
-      );
-    }
-    if (option.handleClick) {
-      return (
-        <MenuOption
-          label={option.label}
-          key={option.id}
-          icon={option.icon}
-          handleClick={option.handleClick}
-        />
-      );
-    }
-    return null;
-  });
-};
-
-const Menu = ({ options, handleClose }: IMenu) => {
+const Menu = (props: IMenu) => {
+  const { options, handleClose } = props;
   const mobileMenuRef = useRef<HTMLDivElement>(null);
 
   const handleWindowClick = useCallback(
@@ -54,7 +22,6 @@ const Menu = ({ options, handleClose }: IMenu) => {
 
   useEffect(() => {
     window.addEventListener("mousedown", handleWindowClick);
-
     return () => {
       window.removeEventListener("mousedown", handleWindowClick);
     };
@@ -63,7 +30,9 @@ const Menu = ({ options, handleClose }: IMenu) => {
   return (
     <StyledMenu ref={mobileMenuRef}>
       <StyledMenuContainer>
-        <Stack direction="column">{renderMenuItems(options)}</Stack>
+        <Stack direction="column">
+          <RenderMenuItems options={options} />
+        </Stack>
       </StyledMenuContainer>
     </StyledMenu>
   );
