+++ src/design/select/OptionItem/styles.ts
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 import { inube } from "@inubekit/inubekit";
+import { basic } from "@design/tokens";
 
 const StyledOptionItemChecked = styled.li`
   display: flex;
@@ -11,11 +12,7 @@ const StyledOptionItemChecked = styled.li`
   border-left-width: 2px;
   border-left-color: ${({ theme }) =>
     theme?.palette?.neutral?.N10 || inube.palette.neutral.N10};
-  padding-top: "4px";
-  padding-right: "16px";
-  padding-bottom: "4px";
-  padding-left: "12px";
-
+  padding: ${basic.spacing.s4} ${basic.spacing.s200};
   p {
     color: ${({ theme }) =>
       theme?.palette?.neutral?.N900 || inube.palette.neutral.N900};
