+++ src/design/select/index.tsx
@@ -1,29 +1,6 @@
 import { useEffect, useRef, useState } from "react";
-import { Size, Status } from "@ptypes/select/selectbusinessUnit/Istatus";
+import { ISelectCheck } from "@ptypes/navigation/ISelectCheck";
 import { SelectCheckUI } from "./interface";
-import { IOptionItemChecked } from "./OptionItem";
-
-interface ISelectCheck {
-  id: string;
-  name: string;
-  options: IOptionItemChecked[];
-  value: string | number;
-
-  disabled?: boolean;
-  fullwidth?: boolean;
-  label?: string;
-  message?: string;
-  onBlur?: (event: FocusEvent) => void;
-  onChange?: (event: React.ChangeEvent<HTMLInputElement>, name: string) => void;
-  onChangeCheck?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  onClick?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  onFocus?: (event: FocusEvent) => void;
-  placeholder?: string;
-  readonly?: boolean;
-  required?: boolean;
-  size?: Size;
-  status?: Status;
-}
 
 const SelectCheck = (props: ISelectCheck) => {
   const {
@@ -31,7 +8,6 @@ const SelectCheck = (props: ISelectCheck) => {
     name,
     options,
     value,
-
     disabled = false,
     fullwidth = false,
     label,
