+++ src/types/navigation/IMenu/index.ts
@@ -1,0 +1,7 @@
+import { IOption } from "@ptypes/navigation/IOption";
+interface IMenu {
+  options: IOption[];
+  handleClose: () => void;
+}
+
+export type { IMenu };
