+++ src/types/navigation/IOptionItemChecked/index.ts
@@ -1,0 +1,8 @@
+interface IOptionItemChecked {
+  id: string;
+  label: string;
+  checked?: boolean;
+  onchange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+}
+
+export type { IOptionItemChecked };
