+++ src/types/navigation/IOptionList/index.ts
@@ -1,0 +1,6 @@
+interface IOptionList {
+  children: React.ReactNode;
+  onClick?: React.ChangeEventHandler<HTMLInputElement>;
+}
+
+export type { IOptionList };
