+++ src/types/navigation/IRenderMenuItems/index.ts
@@ -1,0 +1,6 @@
+import { IOption } from "@ptypes/navigation/IOption";
+interface IRenderMenuItems {
+  options: IOption[];
+}
+
+export type { IRenderMenuItems };
