+++ src/types/navigation/ISelectCheckUI/index.ts
@@ -1,0 +1,8 @@
+import { ISelectCheck } from "../ISelectCheck";
+
+interface ISelectCheckUI extends ISelectCheck {
+  displayList: boolean;
+  focused?: boolean;
+}
+
+export type { ISelectCheckUI };
