+++ src/types/navigation/menuSection/ISection/index.ts
@@ -1,5 +1,4 @@
-import { IMenuItem } from "@design/navigation/MenuItem";
-
+import { IMenuItem } from "@ptypes/navigation/menuItem/IMenuItem";
 interface ISection {
   title?: string;
   links: IMenuItem[];
